Getting an entity from a TOML file now recreates the comments correctly.
